refactor: move HTTP dispatcher test services
